with bugs

+++ Projet2/lectureVilles.cpp
@@ -7,7 +7,7 @@
 
 using namespace std;
 
-void lectureVilles(int popMin, string* &villesNom, int* &villesPop, float* &villesLon, float* &villesLat, int* &villesDeparte, int* &beginDeparte,int &nbVilles){
+void lectureVilles(int popMin, string* &villesNom, int* &villesPop, double* &villesLon, double* &villesLat, int* &villesDeparte, int* &beginDeparte,int &nbVilles){
 
 //-----------------------------------------------------------------
 //--- LECTURE des villes de 'popMin' habitants ou plus
@@ -41,10 +41,10 @@ void lectureVilles(int popMin, string* &villesNom, int* &villesPop, float* &vill
     //villesNom = (string*)_mm_malloc(nbVilles * sizeof(string),VEC_ALIGN);
     //villesPop = new int[nbVilles];
     villesPop = (int*)_mm_malloc(nbVilles * sizeof(int),VEC_ALIGN);
-    //villesLon = new float[nbVilles];
-    villesLon = (float*)_mm_malloc(nbVilles * sizeof(float),VEC_ALIGN);
-    //villesLat = new float[nbVilles];
-    villesLat = (float*)_mm_malloc(nbVilles * sizeof(float),VEC_ALIGN);
+    //villesLon = new double[nbVilles];
+    villesLon = (double*)_mm_malloc(nbVilles * sizeof(double),VEC_ALIGN);
+    //villesLat = new double[nbVilles];
+    villesLat = (double*)_mm_malloc(nbVilles * sizeof(double),VEC_ALIGN);
     villesDeparte = (int*)_mm_malloc(nbVilles * sizeof(int),VEC_ALIGN);
     beginDeparte = (int*) _mm_malloc((NB_DEPART+1) * sizeof(int),VEC_ALIGN); //start at index 0
     // Lecture/Sauvegarde des données
@@ -73,8 +73,8 @@ void lectureVilles(int popMin, string* &villesNom, int* &villesPop, float* &vill
       string::size_type sz;
       string my_nom = ss[0].str();
       int    my_pop = stoi(ss[1].str(),&sz);
-      float  my_lon = stof(ss[2].str(),&sz);
-      float  my_lat = stof(ss[3].str(),&sz);
+      double  my_lon = stof(ss[2].str(),&sz);
+      double  my_lat = stof(ss[3].str(),&sz);
       int    my_dep = stoi(ss[4].str(),&sz);
 
       if(my_pop >= popMin){
